Move visiting traits to type library
